Bootrom running until $15. Memory read and write offsets done

+++ memory.h
@@ -4,8 +4,8 @@
 
 namespace Memory
 {
-        uint8_t Read(uint8_t address);
-        void Write(uint8_t address, uint8_t data);
+        uint8_t Read(uint16_t address);
+        void Write(uint16_t address, uint8_t data);
 
         enum BUS_POINTERS
         {
